+++ test/helper/utils.lua
@@ -25,7 +25,7 @@ local function cleanup(g)
     end
     set_sections(g, { { filename = 'schema.yml', content = box.NULL } })
 
-    g.cluster.main_server.net_box:eval([[require('cartridge').config_patch_clusterwide({migrations = {applied = {}}})]])
+    g.cluster.main_server.net_box:eval([[require('cartridge').config_patch_clusterwide({migrations = {applied = box.NULL }})]])
     local spaces_to_remove = { "first", "sharded" }
     for _, server in ipairs(g.cluster.servers) do
         for _, space in pairs(spaces_to_remove) do
@@ -45,6 +45,35 @@ local function cleanup(g)
                 ]], { space }))
             end)
         end
+
+        -- Cleanup _migrations space.
+        g.cluster:retrying({ timeout = 5 }, function()
+            t.assert(server:eval([[
+                if box.info.ro then
+                    return true
+                end
+                if box.sequence._migrations_id_seq ~= nil then
+                    box.sequence._migrations_id_seq:reset()
+                    box.sequence._migrations_id_seq:set(0)
+                end
+                if box.space._migrations ~= nil and box.space._migrations:truncate() ~= nil then
+                    return false
+                end
+                return true
+            ]]))
+        end)
+        g.cluster:retrying({ timeout = 5 }, function()
+            t.assert(server:eval([[
+                return box.space._migrations:len() == 0
+            ]]))
+        end)
+    end
+
+    -- Reset loader to default.
+    for _, server in pairs(g.cluster.servers) do
+        server:eval([[require('migrator').set_loader(
+            require('migrator.directory-loader').new('test/integration/migrations'))
+        ]])
     end
 
     g.cluster:retrying({ timeout = 1 }, function()